Addressed all comments in PR.
