+++ Testing/boto3_mock.py
@@ -1,5 +1,4 @@
 import json
-from botocore.exceptions import ClientError
 
 
 class Boto3ClientMock():
