+++ Testing/requests_mock.py
@@ -1,12 +1,4 @@
-class MockResponse:
-    def __init__(self, status_code=200, ok=True, reason=None, json_data={}):
-        self.status_code = status_code
-        self.ok = ok
-        self.reason = reason
-        self.json_data = json_data
-
-    def json(self):
-        return self.json_data
+from Testing.util import MockResponse
 
 
 def mock_get(url, headers, params):
